refactor: copy folder in container w/o wildcard
With a wildcard to the COPY command, the container engine would follow
symbolic links and copy their contents to the container, allowing us to
de-duplicate scripts for almost-identical fuzzers (e.g. different
flavours of LibAFLGo).

Unfortunately using a wildcard does not preserve directory structure and
would not allow for subdirectories within a fuzzer's folder. This breaks
the analyses that have code directly in the repository.

This commit reverts the use of the wildcard and uses hard copy instead
of a symbolic links.